requirements.txt
dvc==2.54.0
dvc-s3==2.21.0
scikit-learn==1.2.2
matplotlib==3.7.1
pandas==2.0.0
python-box==7.0.1
xgboost==1.7.5